rename consts - fixed py3 path errors on tests
